pyproject.toml
[tool.poetry]
name = "datavis-app"
version = "0.1.0"
description = "AI-Powered Data Visualization Platform"
authors = ["Votre Nom <votre.email@example.com>"]
readme = "README.md"
packages = [{include = "datavisapp", from = "src"}]


[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.33.0"
pandas = "^2.2.0"
plotly = "^5.18.0"
seaborn = "^0.13.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
openai = "^1.12.0"
numpy = "^1.26.0"
scipy = "^1.12.0"
boto3 = "^1.34.0"
anthropic = "^0.45.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^24.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"